--- configs/parsing_test_1/config.yaml
@@ -1,14 +1,0 @@
-defaults:
-  - dataset/zstruct_monkey
-  - dataset/nwb/monkey_brain
-  - dataset/xpc/monkey_brain
-  - decoder/decoder
-  - trainer/trainer
-  - _self_
-
-dataset:
-  date: 2023-04-03
-  runs: [3]
-  subject: Joker
-  nwb:
-    save: true
