--- configs/parsing_test_1/dataset/nwb/human_emg.yaml
@@ -1,24 +1,0 @@
-compression:
-  type: "gzip"
-  options: 4
-save: true
-institution: "University of Michigan"
-lab: "Chestek Lab"
-behavior_name: "finger_positions"
-behavior_description: "Finger flexion position [thumb, index, middle, ring, pinky], from fully extended (0) to fully flexed (1)"
-behavior_unit: "flexion units"
-behavior_comments: "Raw position data"
-signal_name: "signal"
-signal_description: "Mean absolute value of EMG"
-signal_conversion: 2.5e-7
-signal_filtering: "bandpass: 100-500 Hz"
-signal_comments: "Mean Absolute Value, for each 1ms bin"
-spikes_name: "spikes"
-device: 
-  name: "Cerebus"
-  description: "Data acquisition system"
-  manufacturer: "Blackrock Neurotech"
-electrode_group:
-  name: "emg_electrodes"
-  description: "EMG electrodes"
-  location: "XXX"
