--- configs/parsing_test_1/dataset/zstruct_human.yaml
@@ -1,11 +1,0 @@
-type: zstruct
-date: ???
-runs: ???
-server_path: "Z:/Data/Humans/RPNI"
-subject: ???
-is_monkey: false
-data_path: null
-  
-
-
-
