--- configs/test_linear_decoder/config.yaml
@@ -1,3 +1,0 @@
-defaults:
-  - decoder/testLinearDecoder
-  - _self_
