--- configs/test_new_trainer/config.yaml
@@ -1,4 +1,0 @@
-defaults:
-  - trainer/testTrainer
-  - preprocessing/preprocessing_trainer
-  - _self_
