--- configs/test_trainer/config.yaml
@@ -1,3 +1,0 @@
-defaults:
-  - trainer/testTrainer
-  - _self_
