+++ tests/configs/parsing_test_1/dataset/nwb/monkey_brain.yaml
@@ -1,0 +1,24 @@
+compression:
+  type: "gzip"
+  options: 4
+save: true
+institution: "University of Michigan"
+lab: "Chestek Lab"
+device: 
+  name: "Cerebus"
+  description: "Data acquisition system"
+  manufacturer: "Blackrock Neurotech"
+electrode_group:
+  name: "Utah Array"
+  description: "Utah Array"
+  location: "M1"
+spikes_name: "spikes"
+signal_name: "signal"
+signal_description: "Spiking band power of brain activity"
+signal_comments: "Spiking Band Power (SBP) - i.e. the Mean Absolute Value, for each 1ms bin"
+signal_conversion: 2.5e-7
+signal_filtering: "bandpass: 300-1000 Hz"
+behavior_name: "finger_positions"
+behavior_description: "Finger flexion position [thumb, index, middle, ring, pinky], from fully extended (0) to fully flexed (1)"
+behavior_unit: "flexion units"
+behavior_comments: "Raw position data"
