+++ tests/configs/parsing_test_1/dataset/xpc/human_emg.yaml
@@ -1,0 +1,6 @@
+behavior_field: "FingerAnglesTIMRL"
+signal_field: "EMGFeats"
+experiment_time_field: "ExperimentTime"
+spikes_exist: false
+spikes_field: null
+num_channels: 16
