+++ tests/configs/parsing_test_1/dataset/zstruct_monkey.yaml
@@ -1,0 +1,7 @@
+type: zstruct
+date: ???
+runs: ???
+server_path: "Z:/Data/Monkeys"
+subject: ???
+is_monkey: true
+data_path: null
