+++ tests/configs/test_linear_decoder/config.yaml
@@ -1,0 +1,3 @@
+defaults:
+  - decoder/testLinearDecoder
+  - _self_
