+++ tests/configs/test_new_trainer/config.yaml
@@ -1,0 +1,4 @@
+defaults:
+  - trainer/testTrainer
+  - preprocessing/preprocessing_trainer
+  - _self_
