+++ tests/configs/test_new_trainer/model/model.yaml
