+++ tests/configs/test_trainer/trainer/testTrainer.yaml
@@ -1,0 +1,41 @@
+model:
+  type: LSTM  
+  parameters:
+    input_size: 1
+    num_outputs: 10
+    hidden_size: 64
+    num_layers: 2  
+    rnn_type: lstm  
+    device: cpu  
+    hidden_noise_std: 0.0  
+    dropout_input: False 
+    drop_prob: 0.2  
+
+optimizer:
+  type: Adam
+  params:
+    lr: 0.001
+    weight_decay: 0.0001
+
+scheduler:
+  type: StepLR
+  params:
+    step_size: 10
+    gamma: 0.5
+
+loss_func:
+  type: MSELoss
+  params: {}
+
+training:
+  num_epochs: 2
+  batch_size: 32
+  device: cpu # cuda , if avail 
+  print_results: True
+  print_every: 5
+
+data:
+  train_data:
+    path: "./dataset/train.npz"
+  valid_data:
+    path: "./dataset/valid.npz"
